+++ board/utils.h
@@ -1,3 +1,28 @@
+#define MIN(a,b) \
+ ({ __typeof__ (a) _a = (a); \
+     __typeof__ (b) _b = (b); \
+   (_a < _b) ? _a : _b; })
+
+#define MAX(a,b) \
+ ({ __typeof__ (a) _a = (a); \
+     __typeof__ (b) _b = (b); \
+   (_a > _b) ? _a : _b; })
+
+#define ABS(a) \
+ ({ __typeof__ (a) _a = (a); \
+   (_a > 0) ? _a : (-_a); })
+
+#ifndef NULL
+#define NULL ((void*)0)
+#endif
+
+// STM32 HAL defines this
+#ifndef UNUSED
+#define UNUSED(x) ((void)(x))
+#endif
+
+#define COMPILE_TIME_ASSERT(pred) ((void)sizeof(char[1 - (2 * ((int)(!(pred))))]))
+
 // compute the time elapsed (in microseconds) from 2 counter samples
 // case where ts < ts_last is ok: overflow is properly re-casted into uint32_t
 uint32_t get_ts_elapsed(uint32_t ts, uint32_t ts_last) {
